requirement.txt
schedule
kafka-python
googlefinance
cassandra-driver